Fix checks
